requirements.txt
mysql-connector-python==8.0.32
protobuf==3.20.3
psycopg2==2.9.5
tk==0.1.0
pymongo~=4.3.3
jsonpickle~=3.0.1